Try hacky uint32_t cast to avoid hicpp-signed-bitwise violation

+++ ros1/src/hardware/interface/march_hardware/include/march_hardware/ethercat/imotioncube_pdo_map.h
@@ -28,7 +28,7 @@ struct IMCObject {
         , length(_length)
     {
         uint32_t MSword = (address << 16U);
-        uint32_t LSword = (sub_index << 8U) | length;
+        uint32_t LSword = ((uint32_t)(sub_index << 8U)) | length;
 
         combined_address = (MSword | LSword);
     }
